+++ pkg/wdb/databases.go
@@ -1,7 +1,7 @@
 package wdbClient
 
 import (
-	"github.com/TanmoySG/wunderDB/internal/roles"
+	"github.com/TanmoySG/wunderDB/internal/roles/sysroles"
 	"github.com/TanmoySG/wunderDB/model"
 	er "github.com/TanmoySG/wunderDB/pkg/wdb/errors"
 )
@@ -17,7 +17,7 @@ func (wdb wdbClient) AddDatabase(databaseId model.Identifier, userId model.Ident
 
 	wdb.Databases.CreateDatabase(databaseId, model.Access{})
 
-	return wdb.grantSystemDefaultRole(userId, roles.DatabaseAdminRole, databaseId.String())
+	return wdb.grantSystemDefaultRole(userId, sysroles.DatabaseAdminRole, databaseId.String())
 }
 
 func (wdb wdbClient) GetDatabase(databaseId model.Identifier) (*model.Database, *er.WdbError) {
